requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
pydantic==2.9.2
pymongo==4.10.1
black==24.10.0
paramiko==3.5.0
requests==2.32.3
pyjwt==2.9.0
